move middleware into http component
